Rename `ClientPlayerEntityProxy` to `PlayerUtils`, and move methods under original `PlayerUtils` into it
