+++ common/src/test/java/com/github/khanshoaib3/minecraft_access/test_utils/annotations/FeatureText.java
@@ -18,7 +18,7 @@ import java.lang.annotation.Target;
         MockConfigExtension.class,
         MockMinecraftClientExtension.class,
         MockPlayerPositionUtilsExtension.class,
-        MockClientPlayerEntityProxyExtension.class,
+        MockPlayerUtilsExtension.class,
         MockIntervalExtension.class
 })
 public @interface FeatureText {
