+++ common/src/test/java/com/github/khanshoaib3/minecraft_access/test_utils/annotations/MockPlayerUtils.java
@@ -1,0 +1,18 @@
+package com.github.khanshoaib3.minecraft_access.test_utils.annotations;
+
+import com.github.khanshoaib3.minecraft_access.test_utils.extensions.MockPlayerUtilsExtension;
+import com.github.khanshoaib3.minecraft_access.utils.PlayerUtils;
+import org.mockito.MockedStatic;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Tag for generating {@link MockedStatic} {@link PlayerUtils} instances, see {@link MockPlayerUtilsExtension}.
+ */
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface MockPlayerUtils {
+}
