+++ common/src/test/java/com/github/khanshoaib3/minecraft_access/test_utils/extensions/MockPlayerUtilsExtension.java
@@ -2,8 +2,8 @@ package com.github.khanshoaib3.minecraft_access.test_utils.extensions;
 
 
 import com.github.khanshoaib3.minecraft_access.test_utils.ExtensionUtils;
-import com.github.khanshoaib3.minecraft_access.test_utils.annotations.MockClientPlayerEntityProxy;
-import com.github.khanshoaib3.minecraft_access.utils.ClientPlayerEntityProxy;
+import com.github.khanshoaib3.minecraft_access.test_utils.annotations.MockPlayerUtils;
+import com.github.khanshoaib3.minecraft_access.utils.PlayerUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.AfterEachCallback;
@@ -15,15 +15,15 @@ import java.util.Objects;
 
 /**
  * At {@link BeforeEach} phase,
- * assign new {@link MockedStatic} {@link ClientPlayerEntityProxy} instances to first field that tagged with {@link MockClientPlayerEntityProxy}.
+ * assign new {@link MockedStatic} {@link PlayerUtils} instances to first field that tagged with {@link MockPlayerUtils}.
  * Close the mocked static instance at {@link AfterEach} phase.
  */
-public class MockClientPlayerEntityProxyExtension implements BeforeEachCallback, AfterEachCallback {
-    private MockedStatic<ClientPlayerEntityProxy> ms;
+public class MockPlayerUtilsExtension implements BeforeEachCallback, AfterEachCallback {
+    private MockedStatic<PlayerUtils> ms;
 
     @Override
     public void beforeEach(ExtensionContext extensionContext) {
-        this.ms = ExtensionUtils.mockStaticForAnnotatedField(extensionContext, ClientPlayerEntityProxy.class, MockClientPlayerEntityProxy.class);
+        this.ms = ExtensionUtils.mockStaticForAnnotatedField(extensionContext, PlayerUtils.class, MockPlayerUtils.class);
     }
 
     @Override
